Add lfsr example

+++ generate_notebooks.py
@@ -1,6 +1,7 @@
 import glob
 import subprocess
 import os
+from py2nb.tools import python_to_notebook
 
 with open("README.md", "w") as readme:
     readme.write("""\
@@ -16,11 +17,13 @@ pip install -e .
 """)
     for file in glob.glob("examples/*.py"):
         base_name = file.split("/")[1].split(".")[0]
-        os.remove(f"./notebooks/{base_name}.py")
+        if os.path.isfile(f"./notebooks/{base_name}.py"):
+            os.remove(f"./notebooks/{base_name}.py")
         os.symlink(f"../{file}", f"./notebooks/{base_name}.py")
         readme.write(f"* [{base_name}](./notebooks/{base_name}.ipynb)\n")
 
     for file in glob.glob("examples/*.py"):
         base_name = file.split("/")[1].split(".")[0]
-        subprocess.call(f"python -m py2nb {file} notebooks/{base_name}.ipynb".split(" "))
+
+        python_to_notebook(input_filename=file, output_filename=f"notebooks/{base_name}.ipynb")
         subprocess.call(f"jupyter nbconvert --to notebook --inplace --execute notebooks/{base_name}.ipynb".split(" "))
